requirements.txt
fastapi==0.82.0
uvicorn==0.18.3
google-api-python-client==2.61.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
pytz==2022.2.1